Logic Diagram
This Diagram is used to specify how the relations will be defined in a relational database model, also called Logic Database Diagram. The files are in .PNG and .ERDPLUS format. The tool used to create this diagram is called ERDPlus and is a free open-source tool avaliable in: https://erdplus.com